Photo Saving

+++ Common-JPA/src/main/java/org/example/Entity/AppPhoto.java
@@ -19,5 +19,5 @@ public class AppPhoto {
     private String telegramFileId;
     @OneToOne
     private BinaryContent  binaryContent;
-    private Long fileSize;
+    private Integer fileSize;
 }
